tests to publish and wrapper module
