+++ tests/test_publish_parq.py
@@ -1,92 +1,55 @@
 import pytest
+from mock import patch
 import pandas as pd
 from moto import mock_s3
 from dfmock import DFMock
-from s3_parq import S3Parq
+from s3_parq.publish_parq import S3PublishParq 
 from .mock_helper import MockHelper
 
 @mock_s3
 class Test:
 
-    strings = ("bucket","dataset","key_prefix",)
-    def parq_setup_exclude(self, to_exclude):
+    def publish_parq_setup(self, overrides:dict=dict()):
         df = MockHelper(count=100).dataframe 
         defaults ={
         'bucket':'safebucketname',
         'dataset' : 'safedatasetname',
         'key_prefix' : 'safekeyprefixname',
-        'filters' :{"filtered_thing":"filter_val"},
-        'dataframe' : df
+        'dataframe' : df,
+        'partitions':[]
         }
-        parq = S3Parq()
-        for a in defaults.keys():
-            if a not in to_exclude:
-                setattr(parq,a,defaults[a])
-        return parq
-
-    # requires dataframe
-    def test_requires_dataframe(self):
-        parq = self.parq_setup_exclude(('dataframe',))
-        with pytest.raises(ValueError):
-            parq.publish()
-
-    # only accepts dataframe
-    def test_accepts_only_dataframe(self):
-        with pytest.raises(TypeError):
-            parq = self.parq_setup_exclude(('dataframe',))
-            parq.dataframe = 'not a dataframe'
-
-    # accepts good dataframe
-    def test_accepts_valid_dataframe(self):
-            parq = self.parq_setup_exclude([])
-            parq.publish()
-
-    # accepts good string
-    def test_accepts_valid_dataframe(self):
-        parq = self.parq_setup_exclude([])
-        parq.publish()
-
-    # requires string
-    def test_requires_string(self):
-        for s in self.strings:
-            if s != 'key_prefix':
-                with pytest.raises(ValueError):
-                    parq = self.parq_setup_exclude((s,))
-                    parq.publish()
-
-    # only accepts string
-    def test_accepts_only_string(self):
-        for s in self.strings:
-            parq = self.parq_setup_exclude([])
-            with pytest.raises(TypeError):
-                setattr(parq,s,['this is not a string'])
-
-
-# raises not implemented error for timedelta column types
-    def test_not_implemented_timedelta(self):
-        df = DFMock(count=100)
-        df.columns = {"time":"timedelta","stringer":"string"}
-        df.generate_dataframe()
-        parq = S3Parq()
-        with pytest.raises(NotImplementedError):
-            parq.dataframe = df.dataframe
-
+        return S3PublishParq(   bucket=overrides.get('bucket',defaults['bucket']),
+                                dataset=overrides.get('dataset',defaults['dataset']),
+                                key_prefix=overrides.get('key_prefix',defaults['key_prefix']),
+                                dataframe=overrides.get('dataframe',defaults['dataframe']),
+                                partitions=overrides.get('partitions',defaults['partitions'])
+                            )
 # accepts valid column names as partitions
     def test_accepts_valid_partitions(self):
-        parq = self.parq_setup_exclude([])
-        partitions = parq.dataframe.columns[:1]
-        parq.publish(partitions=partitions)
+        df = DFMock(count=100)
+        df.columns = {"text_col":"string","int_cal":"integer","float_col":"float"}
+        df.generate_dataframe()
+        partitions = df.columns[:1]
+        self.publish_parq_setup(overrides={"dataframe":df.dataframe, "partitions":partitions})
 
-# only accepts valid column names as partitions
+    # only accepts valid column names as partitions
     def test_accepts_valid_partitions(self):
-        parq = self.parq_setup_exclude([])
-        partitions = parq.dataframe.columns[:1]
+        df = DFMock(count=100)
+        df.columns = {"text_col":"string","int_cal":"integer","float_col":"float"}
+        df.generate_dataframe()
+        partitions = df.dataframe.columns[:1]
         partitions += ('banana')
         with pytest.raises(ValueError):
-            parq.publish(partitions=partitions)
-
-# generates partitions in order
-
+            self.publish_parq_setup(overrides={"dataframe":df.dataframe, "partitions":partitions})
+    '''
+    # generates partitions in order
+    @patch('s3_parq.publish_parq.pyarrow.parquet', autospec=True)
+    def test_generates_partitions_in_order(self,mock_parquet):
+        parq = self.parq_setup_exclude([])
+        partitions = parq.dataframe.columns[:1]
+        arg, kwarg = mock_parquet.write_to_dataset.call_args
+        assert kwarg['partition_cols'] == partitions
+    '''
 # - TODO: patch pyarrow here to make sure it is called - that's it.
 # generates parquet filesystem hierarchy correctly
